Get list of certificate and their metadata, or a single certificate

Get list of certificate and their metadata, or a single certificate

as.certificate(x)

certificates(page = 1, per_page = 25, ...)

certificate(id, ...)

certificate_create(name, type, private_key = NULL,
  leaf_certificate = NULL, certificate_chain = NULL,
  dns_names = NULL, ...)

Arguments

x
Object to coerce to an certificate

page
Page to return. Default: 1.

per_page
Number of results per page. Default: 25.

...
Additional arguments passed down to low-level API function (do_*)

id
(numeric) certificate id

name
(character) a certificate name

type
(character) a string representing the type of certificate. The value should be "custom" for a user-uploaded certificate or "lets_encrypt" for one automatically generated with Let's Encrypt. If not provided, "custom" will be assumed by default.

private_key
(character) the contents of a PEM-formatted private-key corresponding to the SSL certificate

leaf_certificate
(character) the contents of a PEM-formatted public SSL certificate

certificate_chain
(character) the full PEM-formatted trust chain between the certificate authority's certificate and your domain's SSL certificate

dns_names
(character) a vector of fully qualified domain names (FQDNs) for which the certificate will be issued. The domains must be managed using DigitalOcean's DNS

Examples

# NOT RUN {
# list certificates
certificates()

# create a certificate (create a fake domain first)
d <- domain_create('tablesandchairsbunnies.stuff', '107.170.220.59')
certificate_create("mycert", "lets_encrypt",
  dns_names = list('tablesandchairsbunnies.stuff'))
# }