Helpers for managing a debian droplets.

Helpers for managing a debian droplets.

debian_add_swap(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

debian_install_r(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE,
  rprofile = "options(repos=c('CRAN'='https://cloud.r-project.org/'))"
)

debian_install_rstudio(
  droplet,
  user = "rstudio",
  password = "server",
  version = "0.99.484",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

debian_install_shiny(
  droplet,
  version = "1.4.0.756",
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE,
  rprofile = "options(repos=c('CRAN'='https://cloud.r-project.org/'))"
)

debian_apt_get_update(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

debian_apt_get_install(
  droplet,
  ...,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

Arguments

droplet

A droplet, or object that can be coerced to a droplet by as.droplet.

user

Default username for Rstudio.

keyfile

Optional private key file.

ssh_passwd

Optional passphrase or callback function for authentication. Refer to the ssh::ssh_connect documentation for more details.

verbose

If TRUE, will print command before executing it.

rprofile

A character string that will be added to the .Rprofile

password

Default password for Rstudio.

version

Version of rstudio to install.

...

Arguments to apt-get install.

Examples

if (FALSE) {
d <- droplet_create()
d %>% debian_add_swap()
d %>% debian_apt_get_update()

d %>% debian_install_r()
d %>% debian_install_rstudio()

# Install libcurl, then build RCurl from source
d %>% debian_apt_get_install("libcurl4-openssl-dev")
d %>% install_r_package("RCurl")
droplet_delete(d)
}