Perform various actions on a droplet.

These droplet actions have no further arguments.

droplet_reboot(droplet, ...)

droplet_power_cycle(droplet, ...)

droplet_shutdown(droplet, ...)

droplet_power_off(droplet, ...)

droplet_power_on(droplet, ...)

droplet_reset_password(droplet, ...)

droplet_enable_ipv6(droplet, ...)

droplet_enable_private_networking(droplet, ...)

droplet_enable_backups(droplet, ...)

droplet_disable_backups(droplet, ...)

droplet_upgrade(droplet, ...)

Arguments

droplet
A droplet, or something that can be coerced to a droplet by as.droplet.

...
Additional options passed down to low-level API method.

Details

reboot

This method allows you to reboot a droplet. This is the preferred method to use if a server is not responding

powercycle

This method allows you to power cycle a droplet. This will turn off the droplet and then turn it back on.

shutdown

Shutdown a running droplet. The droplet will remain in your account and you will continue to be charged for it.

power_off

Shutdown a running droplet. The droplet will remain in your account and you will continue to be charged for it.

reset_password

This method will reset the root password for a droplet. Please be aware that this will reboot the droplet to allow resetting the password.

enable_ipv6

Enable IPv6 networking on an existing droplet (within a region that has IPv6 available).

enable_private_networking

Enable private networking on an existing droplet (within a region that has private networking available)

disable_backups

Disables backups for a droplet.

enable_backups

Enables backups for a droplet.

power_on

Turn on a droplet that's turned off.

Examples

if (FALSE) {
d <- droplets()
d[[1]] %>% droplet_reboot()
d[[2]] %>% droplet_power_cycle()

d <- droplet_create()
d %>% summary
d %>% droplet_enable_backups()
d %>% summary
}