Freeze/thaw droplets.

Freeze powers off the droplet, snapshots to create an image, and deletes the droplet. Thaw performs the inverse: it takes an image and turns it into a running droplet.

droplet_freeze(droplet, name = droplet$name, ...)

droplet_thaw(image, ...)

Arguments

droplet
A droplet, or something that can be coerced to a droplet by as.droplet.

name
Name for the image to be created, or to be used to create a new droplet. Defaults to name of the droplet.

...
For freeze, further args passed on to droplet_snapshot; thaw, args passed on to droplet_create.

image
Image to thaw into a droplet.

Value

droplet_freeze accepts a droplet as first argument, and returns an image; droplet_thaw does the opposite: it accepts an image as first argument, and returns a droplet.

Examples

if (FALSE) {
# freeze
droplet_create(region = 'nyc3') %>% droplet_freeze()

# thaw
droplet_thaw(image='chiromantical-1412718795', region='nyc3')
}