Modify a droplet.

These methods allow you to modify existing droplets.

droplet_resize(droplet, size, ...)

droplet_rebuild(droplet, image, ...)

droplet_rename(droplet, name, ...)

droplet_change_kernel(droplet, kernel, ...)

Arguments

droplet
A droplet, or something that can be coerced to a droplet by as.droplet.

size
(character) Size slug (name) of the image size. See sizes

...
Additional options passed down to low-level API method.

image
(optional) The image ID of the backup image that you would like to restore.

name
(character) The new name for the droplet

kernel
(numeric) The ID of the new kernel.

Details

resize

Resize a specific droplet to a different size. This will affect the number of processors and memory allocated to the droplet.

rebuild

Reinstall a droplet with a default image. This is useful if you want to start again but retain the same IP address for your droplet.

rename

Change the droplet name

change_kernel

Change kernel ID.

Beware: droplet_resize() does not seem to work, see resize()

Examples

if (FALSE) {
droplets()[[1]] %>% droplet_rename(name='newname')
}