Reuse a droplet or image by name

Reuse a droplet or image by name

droplet_reuse(name, ...)

Arguments

name
A name that could be a droplet or image name

...
Named options passed on to droplet_create.

Value

A droplet

Details

Internally, we call the droplets and images (with private = TRUE) to get list of your droplets and images - and we check against those.

Examples

if (FALSE) {
# matches droplet that exists
droplet_reuse(name = 'BeguiledAmmonia')

# matching image that exists
droplet_reuse(name = 'hadleyverse1', size = "1gb")

# no matching droplet or image
droplet_reuse(name = 'tablesandchairs')
}