Take and restore snapshots.

snapshot

Take a snapshot of the droplet once it has been powered off, which can later be restored or used to create a new droplet from the same image.

snapshots_list

List available snapshots

backups_list

List available snapshots

restore

Restore a droplet with a previous image or snapshot. This will be a mirror copy of the image or snapshot to your droplet. Be sure you have backed up any necessary information prior to restore.

droplet_snapshot(droplet, name = NULL, wait = TRUE, ...)

droplet_snapshots_list(droplet, ...)

droplet_restore(droplet, image, ...)

droplet_backups_list(droplet, ...)

Arguments

droplet

A droplet number or the result from a call to droplets()

name

(character) Optional. Name of the new snapshot you want to create. If not set, the snapshot name will default to the current date/time

wait

If TRUE (default), wait until the snapshot has been completed and is ready for use. If set to FALSE we return a droplet object right away after droplet snapshot request has been sent.

...

Additional options passed down to POST

image

(optional) The image ID of the backup image that you would like to restore.

Examples

if (FALSE) {
d <- droplet_create()
d %>% droplet_snapshots_list()
d %>% droplet_backups_list()

d %>%
  droplet_snapshot() %>%
  droplet_power_on() %>%
  droplet_snapshots_list()

# To delete safely
d %>%
  droplet_snapshot() %>%
  droplet_delete() %>%
  action_wait()
}