Helpers for managing a ubuntu droplets.

Helpers for managing a ubuntu droplets.

ubuntu_add_swap(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

ubuntu_install_r(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE,
  rprofile = "options(repos=c('CRAN'='https://cloud.r-project.org/'))"
)

ubuntu_install_rstudio(
  droplet,
  user = "rstudio",
  password = "server",
  version = "0.99.484",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

ubuntu_install_shiny(
  droplet,
  version = "1.4.0.756",
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE,
  rprofile = "options(repos=c('CRAN'='https://cloud.r-project.org/'))"
)

ubuntu_apt_get_cran(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

ubuntu_apt_get_update(
  droplet,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

ubuntu_apt_get_install(
  droplet,
  ...,
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

install_r_package(
  droplet,
  package,
  repo = "https://cloud.r-project.org/",
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

install_github_r_package(
  droplet,
  package,
  repo = "https://cloud.r-project.org/",
  user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

ubuntu_create_user(
  droplet,
  user,
  password,
  ssh_user = "root",
  keyfile = NULL,
  ssh_passwd = NULL,
  verbose = FALSE
)

Arguments

droplet
A droplet, or object that can be coerced to a droplet by as.droplet.

user
Username for non-root account.

keyfile
Optional private key file.

ssh_passwd
Optional passphrase or callback function for authentication. Refer to the ssh::ssh_connect documentation for more details.

verbose
If TRUE, will print command before executing it.

rprofile
A character string that will be added to the .Rprofile

password
Password for non-root acount. Default: root

version
Version of rstudio to install.

...
Arguments to apt-get install.

package
Name of R package to install.

repo
CRAN mirror to use.

ssh_user
(character) User account for ssh commands against droplet.

Examples

if (FALSE) {
d <- droplet_create()
d %>% ubuntu_add_swap()
d %>% ubuntu_apt_get_update()

d %>% ubuntu_install_r()
d %>% ubuntu_install_rstudio()

# Install libcurl, then build RCurl from source
d %>% ubuntu_apt_get_install("libcurl4-openssl-dev")
d %>% install_r_package("RCurl")
droplet_delete(d)
}